+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 
 - Set registry parapheo_url following is_test_url.
   [sgeulette]
+- Override `check_permission` method to `ExternalSessionFeedbackPost` service
+  to ignore `UseRESTAPI` permission and only check the validity of the authentication token.
+  [chris-adam]
 
 1.0b7 (2026-04-02)
 ------------------